+++ src/actions/index.js
@@ -1,2 +1,3 @@
 export * from './warnings'
 export * from './higher-taxonomy-order'
+export * from './round-score'
